+++ Src/Private/Export-Diagrammer.ps1
@@ -232,7 +232,7 @@ function Export-Diagrammer {
                     }
                 }
                 # After adding watermark, convert it to PDF if required. GraphObj is not required for this step.
-                # Needed because .NET does not support adding watermark to PDF files. Used ImageMagick to convert WaterMakerd PNG to PDF.
+                # Needed because .NET does not support adding watermark to PDF files. Used ImageMagick to convert WaterMarked PNG to PDF.
                 if ($Format -eq "pdf") {
                     $Null = ConvertTo-Pdf-WaterMark -ImageInput $Document.FullName -DestinationPath $DestinationPath
                 }